authorization_service/auth-service/requirements.txt
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
fastapi==0.110.1
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dotenv==1.0.1
redis==5.0.3
SQLAlchemy==2.0.29
uvicorn==0.29.0
Werkzeug==3.0.2
orjson==3.10.0
python-multipart==0.0.9
python-jose==3.3.0
cryptography==42.0.5
backoff==2.2.1
requests==2.31.0
email-validator==2.1.1
gunicorn==22.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.8.6
